pipeline_2/requirements.txt
boto3
openpyxl
pandas
redshift_connector
python-dotenv
xlrd
altair
streamlit
pyarrow
fastparquet
psycopg2-binary
